Move the extension FG handler to the PT RootPageService too
